+++ docs/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.4.0 - 2024-09-23
+
+* Binary distributed compiled in Ubuntu 24.04, updated dependencies on the C library -- pls report if you see any issues.
+* Added Erf operator.
+* Added dtypes.MapOfNames that includes its aliases.
+
 # v0.3.2
 
 * Added ReduceAnd and ReduceOr logical operations.